refactor: swap s05/s06 prefixes to match execution order and add pipeline docs
Renamed s05_transforms → s06_transforms and s06_xgboost_classifier →
s05_xgboost_classifier so the sequential numbering reflects actual
execution order (classifier runs before transforms).

Added pipeline execution table to README and onboarding presentation.
Updated outdated project structure slide to reflect src/ layout.
