+++ presentations/onboarding/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "studentprognose-onboarding",
+  "private": true,
+  "scripts": {
+    "dev": "slidev",
+    "build": "slidev build",
+    "export": "slidev export"
+  },
+  "devDependencies": {
+    "@slidev/cli": "^51.0.0",
+    "@slidev/theme-default": "latest"
+  }
+}
