+++ presentations/onboarding/public/chart_distance_histogram.html
@@ -1,0 +1,29 @@
+<!DOCTYPE html>
+<html><head><script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body{margin:0;background:transparent}#chart{width:100%;height:100vh}</style></head>
+<body><div id="chart"></div><script>
+function genDist(n, mean, sd, seed) {
+  const d = []; let r = seed;
+  for (let i = 0; i < n; i++) {
+    r = (r * 16807) % 2147483647;
+    const u1 = r / 2147483647; r = (r * 16807) % 2147483647;
+    const u2 = r / 2147483647;
+    d.push(Math.max(0, mean + sd * Math.sqrt(-2*Math.log(u1)) * Math.cos(2*Math.PI*u2)));
+  }
+  return d;
+}
+const enrolled = genDist(500, 45, 35, 123);
+const notEnrolled = genDist(300, 85, 50, 456);
+const t1 = { x: enrolled, name: 'Ingeschreven', type: 'histogram', opacity: 0.7,
+  marker: { color: '#3b82f6' }, xbins: { size: 15 } };
+const t2 = { x: notEnrolled, name: 'Niet ingeschreven', type: 'histogram', opacity: 0.7,
+  marker: { color: '#ef4444' }, xbins: { size: 15 } };
+const layout = {
+  title: { text: 'Afstand tot universiteit: ingeschreven vs. niet (XGBoost feature)', font: { size: 14 } },
+  xaxis: { title: 'Afstand (km)' }, yaxis: { title: 'Aantal studenten' }, barmode: 'overlay',
+  legend: { orientation: 'h', itemwidth: 40, tracegroupgap: 30, y: 1.12, x: 0.5, xanchor: 'center', font: { size: 10 } },
+  margin: { t: 55, b: 55, l: 50, r: 20 },
+  paper_bgcolor: 'rgba(0,0,0,0)', plot_bgcolor: 'rgba(0,0,0,0)'
+};
+Plotly.newPlot('chart', [t1, t2], layout, { responsive: true, displayModeBar: false });
+</script></body></html>
