+++ presentations/onboarding/public/chart_ensemble_weights_heatmap.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html><head><script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body{margin:0;background:transparent}#chart{width:100%;height:100vh}</style></head>
+<body><div id="chart"></div><script>
+const weeks = ['w1','w5','w10','w15','w17','w20','w25','w30','w35','w38'];
+const models = ['SARIMA Individual','SARIMA Cumulatief','Ratio Model'];
+const z = [
+  [0.30,0.35,0.40,0.45,0.20,0.50,0.55,0.60,0.70,1.00],
+  [0.50,0.45,0.40,0.35,0.80,0.35,0.30,0.25,0.20,0.00],
+  [0.20,0.20,0.20,0.20,0.00,0.15,0.15,0.15,0.10,0.00]
+];
+const data = [{
+  z: z, x: weeks, y: models, type: 'heatmap',
+  colorscale: [[0,'#f0fdf4'],[0.5,'#4ade80'],[1,'#15803d']],
+  text: z.map(row => row.map(v => (v*100).toFixed(0) + '%')),
+  texttemplate: '%{text}', textfont: { size: 11 },
+  hoverongaps: false, showscale: true,
+  colorbar: { title: 'Gewicht', tickformat: '.0%', thickness: 12 }
+}];
+const layout = {
+  title: { text: 'Normal ensemble: gewichten per model per week', font: { size: 14 } },
+  xaxis: { title: 'Week' }, margin: { t: 45, b: 55, l: 120, r: 60 },
+  paper_bgcolor: 'rgba(0,0,0,0)', plot_bgcolor: 'rgba(0,0,0,0)'
+};
+Plotly.newPlot('chart', data, layout, { responsive: true, displayModeBar: false });
+</script></body></html>
