+++ presentations/onboarding/public/chart_examentype_bar.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html><head><script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body{margin:0;background:transparent}#chart{width:100%;height:100vh}</style></head>
+<body><div id="chart"></div><script>
+const types = ['Bachelor', 'Master', 'Pre-master'];
+const actual = [1850, 920, 310];
+const predicted = [1830, 935, 298];
+const t1 = { x: types, y: actual, name: 'Werkelijk', type: 'bar', marker: { color: '#3b82f6' }, width: 0.3, offset: -0.16 };
+const t2 = { x: types, y: predicted, name: 'Voorspeld', type: 'bar', marker: { color: '#16a34a' }, width: 0.3, offset: 0.16 };
+const layout = {
+  title: { text: 'Voorspelling per examentype', font: { size: 15 } },
+  yaxis: { title: 'Aantal studenten' },
+  legend: { orientation: 'h', itemwidth: 40, tracegroupgap: 30, y: 1.1, x: 0.5, xanchor: 'center', font: { size: 11 } },
+  margin: { t: 55, b: 50, l: 55, r: 20 },
+  paper_bgcolor: 'rgba(0,0,0,0)', plot_bgcolor: 'rgba(0,0,0,0)', barmode: 'group'
+};
+Plotly.newPlot('chart', [t1, t2], layout, { responsive: true, displayModeBar: false });
+</script></body></html>
