+++ presentations/onboarding/public/chart_herkomst_trend.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html><head><script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body{margin:0;background:transparent}#chart{width:100%;height:100vh}</style></head>
+<body><div id="chart"></div><script>
+const years = ['2018','2019','2020','2021','2022','2023','2024'];
+const nl = [1700,1750,1620,1800,1850,1870,null];
+const eer = [580,620,540,660,720,760,null];
+const neer = [370,410,390,440,480,490,null];
+const nlP = [null,null,null,null,null,null,1910];
+const eerP = [null,null,null,null,null,null,790];
+const neerP = [null,null,null,null,null,null,510];
+const t1 = { x: years, y: nl, name: 'NL', mode: 'lines+markers', line: { color: '#3b82f6', width: 1.5 }, marker: { size: 4 } };
+const t2 = { x: years, y: eer, name: 'EER', mode: 'lines+markers', line: { color: '#f59e0b', width: 1.5 }, marker: { size: 4 } };
+const t3 = { x: years, y: neer, name: 'Niet-EER', mode: 'lines+markers', line: { color: '#ef4444', width: 1.5 }, marker: { size: 4 } };
+const t4 = { x: ['2023','2024'], y: [1870,1910], name: 'NL voorspeld', mode: 'lines+markers', line: { color: '#3b82f6', width: 1, dash: 'dash' }, marker: { size: 6, symbol: 'star' }, showlegend: false };
+const t5 = { x: ['2023','2024'], y: [760,790], name: 'EER voorspeld', mode: 'lines+markers', line: { color: '#f59e0b', width: 1, dash: 'dash' }, marker: { size: 6, symbol: 'star' }, showlegend: false };
+const t6 = { x: ['2023','2024'], y: [490,510], name: 'Niet-EER voorspeld', mode: 'lines+markers', line: { color: '#ef4444', width: 1, dash: 'dash' }, marker: { size: 6, symbol: 'star' }, showlegend: false };
+const layout = {
+  title: { text: 'Instroom per herkomst over de jaren + voorspelling 2024', font: { size: 14 } },
+  xaxis: { title: 'Collegejaar' }, yaxis: { title: 'Aantal studenten' },
+  legend: { orientation: 'h', itemwidth: 40, tracegroupgap: 30, y: 1.1, x: 0.5, xanchor: 'center', font: { size: 11 } },
+  margin: { t: 55, b: 50, l: 55, r: 20 },
+  paper_bgcolor: 'rgba(0,0,0,0)', plot_bgcolor: 'rgba(0,0,0,0)'
+};
+Plotly.newPlot('chart', [t1, t2, t3, t4, t5, t6], layout, { responsive: true, displayModeBar: false });
+</script></body></html>
