+++ presentations/onboarding/public/chart_mae_per_programme.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html><head><script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body{margin:0;background:transparent}#chart{width:100%;height:100vh}</style></head>
+<body><div id="chart"></div><script>
+const prog = ['Psychologie','Geneeskunde','Rechten','Informatica','Biologie','Bedrijfskunde','Scheikunde','Pedagogiek'];
+const maeInd = [12,8,15,9,7,14,4,8];
+const maeCum = [18,5,12,11,9,10,6,10];
+const maeEns = [8,4,9,7,5,8,3,6];
+const t1 = { x: prog, y: maeInd, name: 'SARIMA Ind.', type: 'bar', marker: { color: '#3b82f6', opacity: 0.8 } };
+const t2 = { x: prog, y: maeCum, name: 'SARIMA Cum.', type: 'bar', marker: { color: '#f59e0b', opacity: 0.8 } };
+const t3 = { x: prog, y: maeEns, name: 'Ensemble', type: 'bar', marker: { color: '#16a34a' } };
+const layout = {
+  title: { text: 'MAE per opleiding per model', font: { size: 14 } },
+  yaxis: { title: 'MAE (studenten)' }, barmode: 'group',
+  legend: { orientation: 'h', itemwidth: 40, tracegroupgap: 30, y: 1.1, x: 0.5, xanchor: 'center', font: { size: 10 } },
+  margin: { t: 55, b: 80, l: 50, r: 20 },
+  paper_bgcolor: 'rgba(0,0,0,0)', plot_bgcolor: 'rgba(0,0,0,0)'
+};
+Plotly.newPlot('chart', [t1, t2, t3], layout, { responsive: true, displayModeBar: false });
+</script></body></html>
