+++ presentations/onboarding/public/chart_model_comparison_radar.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html><head><script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body{margin:0;background:transparent}#chart{width:100%;height:100vh}</style></head>
+<body><div id="chart"></div><script>
+const categories = ['Nauwkeurigheid','Vroege weken','Late weken','Numerus fixus','Kleine opl.','Stabiliteit'];
+const t1 = {
+  type: 'scatterpolar', r: [7, 5, 9, 6, 8, 7, 7], theta: [...categories, categories[0]],
+  fill: 'toself', name: 'SARIMA Individual', fillcolor: 'rgba(59,130,246,0.15)',
+  line: { color: '#3b82f6', width: 1.5 }
+};
+const t2 = {
+  type: 'scatterpolar', r: [7, 7, 6, 8, 5, 8, 7], theta: [...categories, categories[0]],
+  fill: 'toself', name: 'SARIMA Cumulatief', fillcolor: 'rgba(245,158,11,0.15)',
+  line: { color: '#f59e0b', width: 1.5 }
+};
+const t3 = {
+  type: 'scatterpolar', r: [6, 8, 5, 5, 7, 9, 6], theta: [...categories, categories[0]],
+  fill: 'toself', name: 'Ratio Model', fillcolor: 'rgba(168,85,247,0.15)',
+  line: { color: '#a855f7', width: 1.5 }
+};
+const t4 = {
+  type: 'scatterpolar', r: [9, 7, 9, 8, 7, 9, 9], theta: [...categories, categories[0]],
+  fill: 'toself', name: 'Ensemble', fillcolor: 'rgba(22,163,74,0.12)',
+  line: { color: '#16a34a', width: 2 }
+};
+const layout = {
+  title: { text: 'Modelsterkte vergelijking (radar)', font: { size: 14 } },
+  polar: { radialaxis: { visible: true, range: [0, 10] } },
+  legend: { orientation: 'h', itemwidth: 40, tracegroupgap: 30, y: 1.12, x: 0.5, xanchor: 'center', font: { size: 10 } },
+  margin: { t: 65, b: 30, l: 60, r: 60 },
+  paper_bgcolor: 'rgba(0,0,0,0)'
+};
+Plotly.newPlot('chart', [t1, t2, t3, t4], layout, { responsive: true, displayModeBar: false });
+</script></body></html>
