+++ presentations/onboarding/public/chart_numerus_fixus.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html><head><script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body{margin:0;background:transparent}#chart{width:100%;height:100vh}</style></head>
+<body><div id="chart"></div><script>
+const weeks = ['w1','w5','w10','w15','w17','w20','w25','w30','w35','w38'];
+const vooraanmeld = [180,220,280,320,340,355,365,370,372,375];
+const prognose = [350,338,325,318,332,340,335,332,330,330];
+const cap = weeks.map(() => 330);
+const t1 = { x: weeks, y: vooraanmeld, name: 'Vooraanmeldingen', mode: 'lines+markers', line: { color: '#3b82f6', width: 1.5 }, marker: { size: 3 } };
+const t2 = { x: weeks, y: prognose, name: 'Prognose (voor cap)', mode: 'lines+markers', line: { color: '#f59e0b', width: 1.5 }, marker: { size: 3 } };
+const t3 = { x: weeks, y: cap, name: 'Numerus fixus capaciteit', mode: 'lines', line: { color: '#dc2626', width: 1.5, dash: 'dash' } };
+const cappedPrognose = prognose.map(v => Math.min(v, 330));
+const t4 = { x: weeks, y: cappedPrognose, name: 'Prognose (na cap)', mode: 'lines+markers', line: { color: '#16a34a', width: 1.5 }, marker: { size: 4, symbol: 'diamond' } };
+const layout = {
+  title: { text: 'Numerus fixus: voorspelling wordt begrensd op capaciteit', font: { size: 14 } },
+  xaxis: { title: 'Week' }, yaxis: { title: 'Aantal' },
+  legend: { orientation: 'h', itemwidth: 40, tracegroupgap: 30, y: 1.12, x: 0.5, xanchor: 'center', font: { size: 10 } },
+  margin: { t: 55, b: 60, l: 50, r: 20 },
+  paper_bgcolor: 'rgba(0,0,0,0)', plot_bgcolor: 'rgba(0,0,0,0)',
+  shapes: [{ type: 'rect', x0: 'w1', x1: 'w38', y0: 325, y1: 335, fillcolor: 'rgba(220,38,38,0.08)', line: { width: 0 }, layer: 'below' }]
+};
+Plotly.newPlot('chart', [t1, t2, t3, t4], layout, { responsive: true, displayModeBar: false });
+</script></body></html>
