+++ presentations/onboarding/public/chart_residuals.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html><head><script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body{margin:0;background:transparent}#chart{width:100%;height:100vh}</style></head>
+<body><div id="chart"></div><script>
+const prog = ['Psychologie','Geneeskunde','Rechten','Informatica','Biologie','Bedrijfskunde','Scheikunde','Pedagogiek','Sociologie','Filosofie','Tandheelkunde','Geschiedenis'];
+const residuals = [-5, 2, -7, 8, -3, -5, 7, 2, -4, 3, -2, 6];
+const colors = residuals.map(r => r >= 0 ? '#16a34a' : '#dc2626');
+const data = [{
+  y: prog, x: residuals, type: 'bar', orientation: 'h',
+  marker: { color: colors, opacity: 0.8 },
+  text: residuals.map(r => (r >= 0 ? '+' : '') + r), textposition: 'outside', textfont: { size: 11 }
+}];
+const layout = {
+  title: { text: 'Voorspelfout per opleiding (werkelijk - voorspeld)', font: { size: 14 } },
+  xaxis: { title: 'Afwijking (studenten)', zeroline: true, zerolinewidth: 2, zerolinecolor: '#000' },
+  margin: { t: 45, b: 50, l: 110, r: 50 },
+  paper_bgcolor: 'rgba(0,0,0,0)', plot_bgcolor: 'rgba(0,0,0,0)',
+  annotations: [
+    { x: -5, y: 'Rechten', text: 'Overschat', showarrow: false, xshift: -35, font: { size: 9, color: '#dc2626' } },
+    { x: 6, y: 'Informatica', text: 'Onderschat', showarrow: false, xshift: 35, font: { size: 9, color: '#16a34a' } }
+  ]
+};
+Plotly.newPlot('chart', data, layout, { responsive: true, displayModeBar: false });
+</script></body></html>
