+++ presentations/onboarding/public/chart_scatter_pred_vs_actual.html
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html><head><script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body{margin:0;background:transparent}#chart{width:100%;height:100vh}</style></head>
+<body><div id="chart"></div><script>
+let r = 42;
+const points = Array.from({length: 35}, () => {
+  r = (r * 16807) % 2147483647;
+  const actual = 30 + Math.floor((r / 2147483647) * 400);
+  r = (r * 16807) % 2147483647;
+  const noise = ((r / 2147483647) - 0.5) * actual * 0.15;
+  return { actual, predicted: Math.round(actual + noise) };
+});
+const data = [{
+  x: points.map(p => p.actual), y: points.map(p => p.predicted),
+  mode: 'markers', name: 'Opleidingen',
+  marker: { size: 8, color: points.map(p => Math.abs(p.actual - p.predicted)),
+    colorscale: [[0,'#16a34a'],[1,'#dc2626']], showscale: true,
+    colorbar: { title: 'Fout', thickness: 12 }, opacity: 0.8,
+    line: { width: 1, color: '#fff' } },
+  text: points.map((p, i) => 'Opl. ' + (i+1) + '<br>Werkelijk: ' + p.actual + '<br>Voorspeld: ' + p.predicted),
+  hoverinfo: 'text'
+}];
+const perfect = { x: [0, 450], y: [0, 450], mode: 'lines', name: 'Perfecte lijn',
+  line: { color: '#94a3b8', width: 1, dash: 'dash' }, showlegend: true };
+const layout = {
+  title: { text: 'Voorspeld vs. werkelijk per opleiding (scatter)', font: { size: 14 } },
+  xaxis: { title: 'Werkelijk aantal' }, yaxis: { title: 'Voorspeld aantal' },
+  legend: { orientation: 'h', itemwidth: 40, tracegroupgap: 30, y: 1.12, x: 0.5, xanchor: 'center', font: { size: 10 } },
+  margin: { t: 55, b: 55, l: 55, r: 60 },
+  paper_bgcolor: 'rgba(0,0,0,0)', plot_bgcolor: 'rgba(0,0,0,0)'
+};
+Plotly.newPlot('chart', [perfect, ...data], layout, { responsive: true, displayModeBar: false });
+</script></body></html>
