+++ presentations/onboarding/public/chart_waterfall.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html><head><script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body{margin:0;background:transparent}#chart{width:100%;height:100vh}</style></head>
+<body><div id="chart"></div><script>
+const data = [{
+  type: 'waterfall', orientation: 'v',
+  x: ['Vooraanmeldingen','Dubbele aanmeld.','Annuleringen','Geen reactie','Late inschrijvingen','Ingeschreven'],
+  y: [850, -180, -95, -310, 30, 0],
+  measure: ['absolute','relative','relative','relative','relative','total'],
+  text: ['850','-180','-95','-310','+30','295'],
+  textposition: 'outside', textfont: { size: 12 },
+  connector: { line: { color: '#94a3b8', width: 1 } },
+  increasing: { marker: { color: '#16a34a' } },
+  decreasing: { marker: { color: '#ef4444' } },
+  totals: { marker: { color: '#2563eb' } }
+}];
+const layout = {
+  title: { text: 'Van vooraanmelding naar inschrijving (waterfall)', font: { size: 14 } },
+  yaxis: { title: 'Aantal' },
+  margin: { t: 50, b: 60, l: 55, r: 20 },
+  paper_bgcolor: 'rgba(0,0,0,0)', plot_bgcolor: 'rgba(0,0,0,0)'
+};
+Plotly.newPlot('chart', data, layout, { responsive: true, displayModeBar: false });
+</script></body></html>
