+++ presentations/onboarding/public/chart_weekly_applicants_area.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html><head><script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body{margin:0;background:transparent}#chart{width:100%;height:100vh}</style></head>
+<body><div id="chart"></div><script>
+const weeks = Array.from({length: 52}, (_, i) => 'w' + (((i + 38) % 52) + 1));
+const nl = [8,10,14,20,18,15,12,14,16,20,25,22,18,15,12,18,22,26,30,34,42,50,62,72,55,42,32,28,22,18,14,12,10,8,6,5,4,3,2,2,1,1,1,0,0,0,0,0,0,0,0,0];
+const eer = [2,3,4,6,5,4,3,3,4,5,7,6,5,4,3,4,5,6,8,9,12,14,18,22,16,12,8,7,5,4,3,2,2,1,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0];
+const neer = [1,2,2,3,3,2,2,2,3,3,4,3,3,2,2,3,4,5,6,7,9,11,14,18,12,9,6,5,4,3,2,2,1,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0];
+const t1 = { x: weeks, y: nl, name: 'NL', stackgroup: 'one', line: { width: 0.5 }, fillcolor: 'rgba(59,130,246,0.6)' };
+const t2 = { x: weeks, y: eer, name: 'EER', stackgroup: 'one', line: { width: 0.5 }, fillcolor: 'rgba(245,158,11,0.6)' };
+const t3 = { x: weeks, y: neer, name: 'Niet-EER', stackgroup: 'one', line: { width: 0.5 }, fillcolor: 'rgba(239,68,68,0.6)' };
+const layout = {
+  title: { text: 'Wekelijkse aanmeldingen naar herkomst (stacked area)', font: { size: 14 } },
+  xaxis: { title: 'Week', tickangle: -45, dtick: 4 }, yaxis: { title: 'Nieuwe aanmeldingen' },
+  legend: { orientation: 'h', itemwidth: 40, tracegroupgap: 30, y: 1.1, x: 0.5, xanchor: 'center', font: { size: 11 } },
+  margin: { t: 55, b: 70, l: 50, r: 20 },
+  paper_bgcolor: 'rgba(0,0,0,0)', plot_bgcolor: 'rgba(0,0,0,0)'
+};
+Plotly.newPlot('chart', [t1, t2, t3], layout, { responsive: true, displayModeBar: false });
+</script></body></html>
