+++ presentations/onboarding/public/chart_xgb_features.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html><head><script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body{margin:0;background:transparent}#chart{width:100%;height:100vh}</style></head>
+<body><div id="chart"></div><script>
+const features = ['Weeknummer','Examentype','Afstand','Croho opleiding','Herkomst','Numerus fixus','Deadlineweek','Faculteit','Nationaliteit','Geslacht'];
+const importance = [0.18, 0.15, 0.13, 0.12, 0.10, 0.09, 0.08, 0.06, 0.05, 0.04].reverse();
+const colors = importance.map(v => `rgba(124,58,237,${0.3 + v * 4})`);
+const data = [{
+  y: features.slice().reverse(), x: importance, type: 'bar', orientation: 'h',
+  marker: { color: colors }, text: importance.map(v => (v*100).toFixed(0) + '%'), textposition: 'outside'
+}];
+const layout = {
+  title: { text: 'XGBoost Classifier: feature importance', font: { size: 15 } },
+  xaxis: { title: 'Relatief belang', tickformat: '.0%' },
+  margin: { t: 45, b: 50, l: 120, r: 50 },
+  paper_bgcolor: 'rgba(0,0,0,0)', plot_bgcolor: 'rgba(0,0,0,0)'
+};
+Plotly.newPlot('chart', data, layout, { responsive: true, displayModeBar: false });
+</script></body></html>
