+++ presentations/onboarding/public/chart_xgb_probability.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html><head><script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body{margin:0;background:transparent}#chart{width:100%;height:100vh}</style></head>
+<body><div id="chart"></div><script>
+function genProbs(n, mean, seed) {
+  const d = []; let r = seed;
+  for (let i = 0; i < n; i++) {
+    r = (r * 16807) % 2147483647;
+    d.push(Math.min(1, Math.max(0, mean + ((r / 2147483647) - 0.5) * 0.6)));
+  }
+  return d;
+}
+const enrolled = genProbs(400, 0.72, 111);
+const notEnrolled = genProbs(250, 0.35, 222);
+const t1 = { x: enrolled, name: 'Daadwerkelijk ingeschreven', type: 'histogram', opacity: 0.7,
+  marker: { color: '#16a34a' }, xbins: { start: 0, end: 1, size: 0.05 } };
+const t2 = { x: notEnrolled, name: 'Niet ingeschreven', type: 'histogram', opacity: 0.7,
+  marker: { color: '#ef4444' }, xbins: { start: 0, end: 1, size: 0.05 } };
+const layout = {
+  title: { text: 'XGBoost Classifier: verdeling voorspelde kansen', font: { size: 14 } },
+  xaxis: { title: 'Voorspelde inschrijfkans', tickformat: '.0%' },
+  yaxis: { title: 'Aantal studenten' }, barmode: 'overlay',
+  legend: { orientation: 'h', itemwidth: 40, tracegroupgap: 30, y: 1.12, x: 0.5, xanchor: 'center', font: { size: 10 } },
+  margin: { t: 55, b: 55, l: 50, r: 20 },
+  paper_bgcolor: 'rgba(0,0,0,0)', plot_bgcolor: 'rgba(0,0,0,0)',
+  shapes: [{ type: 'line', x0: 0.5, x1: 0.5, y0: 0, y1: 1, yref: 'paper',
+    line: { color: '#000', width: 1, dash: 'dash' } }],
+  annotations: [{ x: 0.5, y: 1.02, yref: 'paper', text: 'Drempelwaarde 50%',
+    showarrow: false, font: { size: 10 } }]
+};
+Plotly.newPlot('chart', [t1, t2], layout, { responsive: true, displayModeBar: false });
+</script></body></html>
