+++ presentations/onboarding/public/chart_xgb_reg_features.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html><head><script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body{margin:0;background:transparent}#chart{width:100%;height:100vh}</style></head>
+<body><div id="chart"></div><script>
+const features = ['Week 38 (laatste)','Week 37','Week 36','Week 20 (deadline)','Week 17','Examentype','Croho opleiding','Herkomst','Collegejaar','Faculteit'];
+const importance = [0.22, 0.14, 0.11, 0.10, 0.09, 0.08, 0.08, 0.07, 0.06, 0.05].reverse();
+const colors = importance.map(v => `rgba(22,163,74,${0.3 + v * 3.5})`);
+const data = [{
+  y: features.slice().reverse(), x: importance, type: 'bar', orientation: 'h',
+  marker: { color: colors }, text: importance.map(v => (v*100).toFixed(0) + '%'), textposition: 'outside'
+}];
+const layout = {
+  title: { text: 'XGBoost Regressor: feature importance', font: { size: 15 } },
+  xaxis: { title: 'Relatief belang', tickformat: '.0%' },
+  margin: { t: 45, b: 50, l: 140, r: 50 },
+  paper_bgcolor: 'rgba(0,0,0,0)', plot_bgcolor: 'rgba(0,0,0,0)'
+};
+Plotly.newPlot('chart', data, layout, { responsive: true, displayModeBar: false });
+</script></body></html>
