+++ presentations/onboarding/public/chart_year_over_year.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html><head><script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body{margin:0;background:transparent}#chart{width:100%;height:100vh}</style></head>
+<body><div id="chart"></div><script>
+const years = ['2018','2019','2020','2021','2022','2023','2024'];
+const enrolled = [2650, 2780, 2550, 2900, 3050, 3120, null];
+const predicted = [null, null, null, null, null, null, 3180];
+const predAll = [2620, 2750, 2600, 2870, 3020, 3090, 3180];
+const t1 = { x: years, y: enrolled, name: 'Werkelijk ingeschreven', type: 'bar',
+  marker: { color: '#3b82f6' }, width: 0.5 };
+const t2 = { x: years, y: predicted, name: 'Voorspeld (2024)', type: 'bar',
+  marker: { color: '#16a34a', pattern: { shape: '/' } }, width: 0.5 };
+const t3 = { x: years, y: predAll, name: 'Historische voorspellingen', mode: 'lines+markers',
+  line: { color: '#f59e0b', width: 1.5, dash: 'dot' }, marker: { size: 4 } };
+const layout = {
+  title: { text: 'Totale eerstejaars instroom per collegejaar', font: { size: 14 } },
+  yaxis: { title: 'Aantal studenten', range: [2200, 3400] },
+  legend: { orientation: 'h', itemwidth: 40, tracegroupgap: 30, y: 1.1, x: 0.5, xanchor: 'center', font: { size: 10 } },
+  margin: { t: 55, b: 50, l: 55, r: 20 }, barmode: 'overlay',
+  paper_bgcolor: 'rgba(0,0,0,0)', plot_bgcolor: 'rgba(0,0,0,0)'
+};
+Plotly.newPlot('chart', [t1, t2, t3], layout, { responsive: true, displayModeBar: false });
+</script></body></html>
