+++ presentations/onboarding/public/cumulative_visual.html
@@ -1,0 +1,120 @@
+<!DOCTYPE html>
+<html>
+<head>
+<script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>
+  body { margin: 0; background: transparent; }
+  #chart { width: 100%; height: 100vh; }
+</style>
+</head>
+<body>
+<div id="chart"></div>
+<script>
+const weeks = [
+  'w39','w40','w41','w42','w43','w44','w45','w46','w47','w48','w49','w50',
+  'w51','w52','w1','w2','w3','w4','w5','w6','w7','w8','w9','w10',
+  'w11','w12','w13','w14','w15','w16','w17','w18','w19','w20',
+  'w21','w22','w23','w24','w25','w26','w27','w28','w29','w30',
+  'w31','w32','w33','w34','w35','w36','w37','w38'
+];
+
+// Actual cumulative data (w39 to w10)
+const actualWeeks = weeks.slice(0, 24);
+const actualData = [5,10,18,28,40,55,68,80,92,105,118,130,142,155,165,178,190,205,218,230,242,255,268,280];
+
+// SARIMA extrapolation (w10 to w38) - continues from last actual point
+const sarimaWeeks = weeks.slice(23);
+const sarimaData = [280,292,303,314,324,333,341,349,356,362,368,373,378,383,387,391,395,398,401,404,407,409,411,413,415,417,418,419];
+
+// XGBoost final prediction
+const xgboostWeek = ['w38'];
+const xgboostValue = [295];
+
+const traceActual = {
+  x: actualWeeks,
+  y: actualData,
+  mode: 'lines+markers',
+  name: 'Werkelijke data',
+  line: { color: '#2563eb', width: 1.5 },
+  marker: { size: 3 }
+};
+
+const traceSarima = {
+  x: sarimaWeeks,
+  y: sarimaData,
+  mode: 'lines+markers',
+  name: 'SARIMA extrapolatie',
+  line: { color: '#f59e0b', width: 1.5, dash: 'dash' },
+  marker: { size: 3 }
+};
+
+const traceXgboost = {
+  x: xgboostWeek,
+  y: xgboostValue,
+  mode: 'markers+text',
+  name: 'XGBoost voorspelling (inschrijvingen)',
+  marker: { color: '#16a34a', size: 16, symbol: 'star' },
+  text: ['295 studenten'],
+  textposition: 'bottom center',
+  textfont: { size: 13, color: '#16a34a', family: 'Arial Black' }
+};
+
+// Annotation area for SARIMA zone
+const layout = {
+  title: { text: 'Cumulatief spoor: van vooraanmeldingen naar voorspelling', font: { size: 18 } },
+  xaxis: {
+    title: 'Week (academisch jaar)',
+    tickangle: -45,
+    dtick: 2
+  },
+  yaxis: { title: 'Aantal' },
+  legend: { orientation: 'h', itemwidth: 40, tracegroupgap: 30, y: 1.12, x: 0.5, xanchor: 'center', font: { size: 10 } },
+  margin: { t: 65, b: 60, l: 60, r: 30 },
+  paper_bgcolor: 'rgba(0,0,0,0)',
+  plot_bgcolor: 'rgba(0,0,0,0)',
+  shapes: [
+    {
+      type: 'rect',
+      x0: 'w10', x1: 'w38',
+      y0: 0, y1: 430,
+      fillcolor: '#fef3c7',
+      opacity: 0.3,
+      line: { width: 0 },
+      layer: 'below'
+    }
+  ],
+  annotations: [
+    {
+      x: 'w5',
+      y: 240,
+      text: '<b>Data bekend</b>',
+      showarrow: false,
+      font: { size: 13, color: '#2563eb' }
+    },
+    {
+      x: 'w24',
+      y: 380,
+      text: '<b>SARIMA extrapoleert</b>',
+      showarrow: false,
+      font: { size: 13, color: '#f59e0b' }
+    },
+    {
+      x: 'w34',
+      y: 250,
+      text: '<b>XGBoost vertaalt<br>vooraanmeldingen<br>naar inschrijvingen</b>',
+      showarrow: true,
+      arrowhead: 2,
+      ax: 0,
+      ay: -40,
+      font: { size: 12, color: '#16a34a' }
+    }
+  ]
+};
+
+Plotly.newPlot('chart', [traceActual, traceSarima, traceXgboost], layout, {
+  responsive: true,
+  displayModeBar: false
+});
+</script>
+</body>
+</html>
