+++ presentations/onboarding/public/ensemble_visual.html
@@ -1,0 +1,70 @@
+<!DOCTYPE html>
+<html>
+<head>
+<script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body { margin: 0; background: transparent; } #chart { width: 100%; height: 100vh; }</style>
+</head>
+<body>
+<div id="chart"></div>
+<script>
+const weeks = ['w1','w2','w3','w4','w5','w6','w7','w8','w9','w10',
+  'w11','w12','w13','w14','w15','w16','w17','w18','w19','w20',
+  'w21','w22','w23','w24','w25','w26','w27','w28','w29','w30',
+  'w31','w32','w33','w34','w35','w36','w37','w38'];
+
+const actual = 295;
+
+// Three model predictions converging over time
+const sarimaInd = [340,335,328,322,318,315,312,310,308,306,304,303,302,301,300,299,298,298,297,297,296,296,296,295,295,295,295,295,295,295,295,295,295,295,295,295,295,295];
+const sarimaCum = [260,265,270,274,278,280,282,284,286,287,288,289,290,291,292,293,293,294,294,294,295,295,295,295,295,295,295,295,295,295,295,295,295,295,295,295,295,295];
+const ratio =     [310,308,306,305,304,303,302,301,300,300,299,299,298,298,298,297,297,297,296,296,296,296,296,295,295,295,295,295,295,295,295,295,295,295,295,295,295,295];
+
+// Ensemble: weighted average that converges fastest
+const ensemble = weeks.map((_, i) => {
+  const w1 = 0.4, w2 = 0.35, w3 = 0.25;
+  return Math.round(sarimaInd[i] * w1 + sarimaCum[i] * w2 + ratio[i] * w3);
+});
+
+const traceInd = {
+  x: weeks, y: sarimaInd, mode: 'lines', name: 'SARIMA Individual',
+  line: { color: '#3b82f6', width: 1, dash: 'dot' }
+};
+const traceCum = {
+  x: weeks, y: sarimaCum, mode: 'lines', name: 'SARIMA Cumulatief',
+  line: { color: '#f59e0b', width: 1, dash: 'dot' }
+};
+const traceRatio = {
+  x: weeks, y: ratio, mode: 'lines', name: 'Ratio Model',
+  line: { color: '#a855f7', width: 1, dash: 'dot' }
+};
+const traceEnsemble = {
+  x: weeks, y: ensemble, mode: 'lines', name: 'Ensemble',
+  line: { color: '#16a34a', width: 2.5 }
+};
+const traceActual = {
+  x: ['w38'], y: [actual], mode: 'markers', name: 'Werkelijk aantal',
+  marker: { color: '#dc2626', size: 12, symbol: 'diamond' }
+};
+
+const layout = {
+  title: { text: 'Ensemble: modellen convergeren naar werkelijk aantal', font: { size: 16 } },
+  xaxis: { title: 'Week', tickangle: -45, dtick: 2 },
+  yaxis: { title: 'Voorspeld aantal studenten', range: [250, 350] },
+  legend: { orientation: 'h', itemwidth: 40, tracegroupgap: 30, y: 1.12, x: 0.5, xanchor: 'center', font: { size: 10 } },
+  margin: { t: 65, b: 60, l: 55, r: 30 },
+  paper_bgcolor: 'rgba(0,0,0,0)',
+  plot_bgcolor: 'rgba(0,0,0,0)',
+  shapes: [{
+    type: 'line', x0: 'w1', x1: 'w38', y0: actual, y1: actual,
+    line: { color: '#dc2626', width: 1, dash: 'dash' }
+  }],
+  annotations: [{
+    x: 'w5', y: actual + 3, text: 'Werkelijk: 295', showarrow: false,
+    font: { size: 11, color: '#dc2626' }
+  }]
+};
+
+Plotly.newPlot('chart', [traceInd, traceCum, traceRatio, traceEnsemble, traceActual], layout, { responsive: true, displayModeBar: false });
+</script>
+</body>
+</html>
