+++ presentations/onboarding/public/individual_visual.html
@@ -1,0 +1,75 @@
+<!DOCTYPE html>
+<html>
+<head>
+<script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body { margin: 0; background: transparent; } #chart { width: 100%; height: 100vh; }</style>
+</head>
+<body>
+<div id="chart"></div>
+<script>
+const weeks = [
+  'w39','w40','w41','w42','w43','w44','w45','w46','w47','w48','w49','w50',
+  'w51','w52','w1','w2','w3','w4','w5','w6','w7','w8','w9','w10',
+  'w11','w12','w13','w14','w15','w16','w17','w18','w19','w20',
+  'w21','w22','w23','w24','w25','w26','w27','w28','w29','w30',
+  'w31','w32','w33','w34','w35','w36','w37','w38'
+];
+
+const applicants = [12,18,22,30,25,20,15,18,22,28,35,30,25,20,15,22,28,32,38,42,55,65,80,95,70,55,40,35,28,22,18,15,12,10,8,6,5,4,3,3,2,2,1,1,1,0,0,0,0,0];
+const predicted = applicants.map((a, i) => {
+  const baseRate = 0.35 + (i / weeks.length) * 0.3;
+  return Math.round(a * baseRate);
+});
+
+const cumPredicted = [];
+let sum = 0;
+for (const p of predicted.slice(0, 24)) { sum += p; cumPredicted.push(sum); }
+
+const sarimaWeeks = weeks.slice(23);
+const sarimaData = [sum];
+for (let i = 1; i < sarimaWeeks.length; i++) {
+  const remaining = predicted[23 + i] || Math.max(0, predicted[23] * Math.exp(-i * 0.15));
+  sum += Math.round(remaining);
+  sarimaData.push(sum);
+}
+
+const traceApplicants = {
+  x: weeks.slice(0, 40), y: applicants.slice(0, 40),
+  mode: 'lines', name: 'Aanmeldingen/week',
+  line: { color: '#94a3b8', width: 1 },
+  fill: 'tozeroy', fillcolor: 'rgba(148,163,184,0.15)', yaxis: 'y2'
+};
+const traceCumPredicted = {
+  x: weeks.slice(0, 24), y: cumPredicted,
+  mode: 'lines+markers', name: 'XGBoost kans → cumulatief',
+  line: { color: '#7c3aed', width: 1.5 }, marker: { size: 2 }
+};
+const traceSarima = {
+  x: sarimaWeeks, y: sarimaData,
+  mode: 'lines', name: 'SARIMA extrapolatie',
+  line: { color: '#f59e0b', width: 1.5, dash: 'dash' }
+};
+
+const layout = {
+  title: { text: 'Individueel: XGBoost voorspelt kans per student → SARIMA extrapoleert', font: { size: 14 } },
+  xaxis: { title: 'Week', tickangle: -45, dtick: 4 },
+  yaxis: { title: 'Cum. inschrijvingen' },
+  yaxis2: { title: 'Aanmeldingen', side: 'right', overlaying: 'y', showgrid: false },
+  legend: { orientation: 'h', itemwidth: 40, tracegroupgap: 30, y: 1.12, x: 0.5, xanchor: 'center', font: { size: 10 } },
+  margin: { t: 60, b: 70, l: 50, r: 50 },
+  paper_bgcolor: 'rgba(0,0,0,0)', plot_bgcolor: 'rgba(0,0,0,0)',
+  shapes: [{
+    type: 'line', x0: 'w10', x1: 'w10', y0: 0, y1: 1, yref: 'paper',
+    line: { color: '#dc2626', width: 1, dash: 'dot' }
+  }],
+  annotations: [
+    { x: 'w10', y: 1.02, yref: 'paper', text: '<b>Vandaag (w10)</b>', showarrow: false, font: { size: 10, color: '#dc2626' } },
+    { x: 'w5', y: cumPredicted[18], text: 'XGBoost: kans<br>per student', showarrow: true, ax: -40, ay: -25, font: { size: 10, color: '#7c3aed' } },
+    { x: 'w25', y: sarimaData[14], text: 'SARIMA vult<br>toekomst aan', showarrow: true, ax: 40, ay: -20, font: { size: 10, color: '#f59e0b' } }
+  ]
+};
+
+Plotly.newPlot('chart', [traceApplicants, traceCumPredicted, traceSarima], layout, { responsive: true, displayModeBar: false });
+</script>
+</body>
+</html>
