+++ presentations/onboarding/public/multiyear_visual.html
@@ -1,0 +1,60 @@
+<!DOCTYPE html>
+<html>
+<head>
+<script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body { margin: 0; background: transparent; } #chart { width: 100%; height: 100vh; }</style>
+</head>
+<body>
+<div id="chart"></div>
+<script>
+const weeks = Array.from({length: 52}, (_, i) => 'w' + (((i + 38) % 52) + 1));
+
+function generateYear(base, seed) {
+  const data = []; let cum = 0;
+  let r = seed;
+  for (let i = 0; i < 52; i++) {
+    r = (r * 16807) % 2147483647;
+    const noise = ((r / 2147483647) - 0.5) * 3;
+    cum += Math.max(0, base * Math.exp(-((i - 20) ** 2) / 300) + noise + i * 0.08);
+    data.push(Math.round(cum));
+  }
+  return data;
+}
+
+const y2021 = generateYear(18, 111); const y2022 = generateYear(20, 222);
+const y2023 = generateYear(22, 333); const y2024full = generateYear(24, 444);
+const y2024partial = y2024full.slice(0, 24);
+const y2024sarima = y2024full.slice(23).map((v, i) => Math.round(v * (1 + (Math.sin(i/5) * 0.02))));
+const xgboostPoint = Math.round(y2024sarima[y2024sarima.length - 1] * 0.72);
+
+const traces = [
+  { x: weeks, y: y2021, name: '2021', line: { color: '#cbd5e1', width: 1 } },
+  { x: weeks, y: y2022, name: '2022', line: { color: '#94a3b8', width: 1 } },
+  { x: weeks, y: y2023, name: '2023', line: { color: '#64748b', width: 1 } },
+  { x: weeks.slice(0, 24), y: y2024partial, name: '2024 (bekend)', line: { color: '#2563eb', width: 1.5 } },
+  { x: weeks.slice(23), y: y2024sarima, name: '2024 (SARIMA)', line: { color: '#f59e0b', width: 1.5, dash: 'dash' } },
+  { x: ['w38'], y: [xgboostPoint], name: 'XGBoost eindvoorspelling', mode: 'markers',
+    marker: { color: '#16a34a', size: 12, symbol: 'star' } },
+].map(t => ({ mode: 'lines', ...t }));
+
+const layout = {
+  title: { text: 'SARIMA leert seizoenspatronen → XGBoost vertaalt naar inschrijvingen', font: { size: 14 } },
+  xaxis: { title: 'Week (academisch jaar)', tickangle: -45, dtick: 4 },
+  yaxis: { title: 'Cumulatieve vooraanmeldingen' },
+  legend: { orientation: 'h', itemwidth: 40, tracegroupgap: 30, y: 1.15, x: 0.5, xanchor: 'center', font: { size: 10 }, traceorder: 'normal' },
+  margin: { t: 65, b: 70, l: 50, r: 30 },
+  paper_bgcolor: 'rgba(0,0,0,0)', plot_bgcolor: 'rgba(0,0,0,0)',
+  shapes: [{
+    type: 'rect', x0: weeks[23], x1: weeks[51], y0: 0, y1: 1, yref: 'paper',
+    fillcolor: '#fef3c7', opacity: 0.2, line: { width: 0 }, layer: 'below'
+  }],
+  annotations: [
+    { x: 'w38', y: xgboostPoint, text: '<b>XGBoost: ' + xgboostPoint + ' studenten</b>', showarrow: true, ax: -60, ay: 25, font: { size: 11, color: '#16a34a' } },
+    { x: weeks[35], y: 0.88, yref: 'paper', text: 'SARIMA extrapoleert', showarrow: false, font: { size: 10, color: '#92400e' }, bgcolor: '#fef3c7', borderpad: 3 }
+  ]
+};
+
+Plotly.newPlot('chart', traces, layout, { responsive: true, displayModeBar: false });
+</script>
+</body>
+</html>
