+++ presentations/onboarding/public/programmes_visual.html
@@ -1,0 +1,58 @@
+<!DOCTYPE html>
+<html>
+<head>
+<script src="https://cdn.plot.ly/plotly-2.35.0.min.js"></script>
+<style>body { margin: 0; background: transparent; } #chart { width: 100%; height: 100vh; }</style>
+</head>
+<body>
+<div id="chart"></div>
+<script>
+const programmes = ['Psychologie','Geneeskunde','Rechten','Informatica','Biologie','Bedrijfskunde','Scheikunde','Tandheelkunde'];
+const actual =     [420, 330, 285, 210, 165, 310, 95, 120];
+const predicted =  [415, 328, 278, 218, 158, 305, 102, 118];
+const capacity =   [null, 330, null, null, null, null, null, 120];
+
+const traceActual = {
+  x: programmes, y: actual, type: 'bar', name: 'Werkelijk',
+  marker: { color: '#3b82f6', opacity: 0.8, line: { width: 0 } },
+  width: 0.35, offset: -0.18
+};
+
+const tracePredicted = {
+  x: programmes, y: predicted, type: 'bar', name: 'Voorspeld (ensemble)',
+  marker: { color: '#16a34a', opacity: 0.8, line: { width: 0 } },
+  width: 0.35, offset: 0.18
+};
+
+const traceCap = {
+  x: programmes.filter((_, i) => capacity[i] !== null),
+  y: capacity.filter(c => c !== null),
+  mode: 'markers', name: 'Numerus fixus capaciteit',
+  marker: { color: '#dc2626', size: 10, symbol: 'line-ew-open', line: { width: 2.5, color: '#dc2626' } }
+};
+
+// Error annotations
+const errors = programmes.map((p, i) => {
+  const diff = predicted[i] - actual[i];
+  const pct = ((Math.abs(diff) / actual[i]) * 100).toFixed(1);
+  return { x: p, y: Math.max(actual[i], predicted[i]) + 12, text: (diff >= 0 ? '+' : '') + diff, showarrow: false,
+    font: { size: 10, color: Math.abs(diff) <= 10 ? '#16a34a' : '#f59e0b' } };
+});
+
+const layout = {
+  title: { text: 'Voorspelling vs. werkelijk per opleiding (voorbeeld)', font: { size: 15 } },
+  xaxis: { title: '', tickangle: -25 },
+  yaxis: { title: 'Aantal studenten' },
+  legend: { orientation: 'h', itemwidth: 40, tracegroupgap: 30, y: 1.12, x: 0.5, xanchor: 'center', font: { size: 10 } },
+  margin: { t: 65, b: 55, l: 55, r: 30 },
+  paper_bgcolor: 'rgba(0,0,0,0)',
+  plot_bgcolor: 'rgba(0,0,0,0)',
+  barmode: 'group',
+  bargap: 0.3,
+  annotations: errors
+};
+
+Plotly.newPlot('chart', [traceActual, tracePredicted, traceCap], layout, { responsive: true, displayModeBar: false });
+</script>
+</body>
+</html>
