+++ src/output/s10_postprocessor.py
@@ -7,7 +7,7 @@ from src.utils.weeks import (
     DataOption, StudentYearPrediction, increment_week, convert_nan_to_zero,
 )
 from src.models.s09_ratio import predict_with_ratio as _predict_with_ratio
-from src.data.s05_transforms import replace_latest_data
+from src.data.s06_transforms import replace_latest_data
 
 
 class PostProcessor:
